Task 02 was finished

+++ src/Tasks/Task01.java
@@ -2,7 +2,7 @@ package Tasks;
 
 public class Task01 {
     public static void main(String[] args) {
-        double x = 5, y = 3;
+        int x = 5, y = 3;
         System.out.println(x / y);
         System.out.println(x % y);
         System.out.println(Math.sqrt(x));
